refactor: extract benchmark ledger case analysis CLI command
