+++ loopx/cli_commands/__init__.py
@@ -60,6 +60,10 @@ from .benchmark_run_ledger import (
     handle_benchmark_run_ledger_command,
     register_benchmark_run_ledger_commands,
 )
+from .benchmark_run_ledger_case_analysis import (
+    handle_benchmark_run_ledger_case_analysis_command,
+    register_benchmark_run_ledger_case_analysis_commands,
+)
 from .benchmark_run_ledger_maintenance import (
     handle_benchmark_run_ledger_maintenance_command,
     register_benchmark_run_ledger_maintenance_commands,
@@ -159,6 +163,7 @@ __all__ = [
     "handle_benchmark_command",
     "handle_benchmark_review_lifecycle_command",
     "handle_benchmark_run_ledger_command",
+    "handle_benchmark_run_ledger_case_analysis_command",
     "handle_benchmark_run_ledger_maintenance_command",
     "handle_bootstrap_connect_command",
     "handle_check_command",
@@ -214,6 +219,7 @@ __all__ = [
     "register_benchmark_command_group",
     "register_benchmark_review_lifecycle_commands",
     "register_benchmark_run_ledger_commands",
+    "register_benchmark_run_ledger_case_analysis_commands",
     "register_benchmark_run_ledger_maintenance_commands",
     "register_bootstrap_connect_command",
     "register_doctor_command",
